Move CalendarDef code
