+++ src/lib/Config.js
@@ -3,7 +3,7 @@
 import _ from 'lodash';
 import Moment from "moment";
 // import * as CalendarsDef from '../calendars';
-import * as CalendarsDef from '../calendars/indexYaml';
+import * as CalendarsDef from './CalendarDef';
 
 /**
  * Config Class
