+++ src/lib/DateItem.js
@@ -4,7 +4,7 @@ import Moment from "moment";
 import {Cycles, LiturgicalSeasons, Types} from "../constants";
 import * as Dates from "./Dates";
 import Sanctoral from "./Sanctoral";
-import * as CalendarsDef from "../calendars/indexYaml";
+import * as CalendarsDef from "./CalendarDef";
 
 export opaque type ISOString: string = string;
